% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_smartTable.R
\name{create_smartTable}
\alias{create_smartTable}
\title{Create a look up table}
\usage{
create_smartTable()
}
\value{
A data base connection to the SQLite database containing the smartTable.
}
\description{
\code{create_smartTable} will create a look up table for random 
field analysis inside an SQLite database. The table is automatically filled each time 
\code{UPMASK} calls the function \code{\link{analyse_randomKde2d_smart}}.
}
\examples{
# Create the table
library(RSQLite)
stcon <- create_smartTable() 

# Clean the environment
dbDisconnect(stcon)

}
\author{
Alberto Krone-Martins, Andre Moitinho
}
\keyword{utilities}
